MB-32164: Log slow/hung kv_sender operations

+++ secondary/common/mutation_test.go
@@ -25,8 +25,8 @@ func TestPayloadVbmap(t *testing.T) {
 
 func TestKVEqual(t *testing.T) {
 	seqno, docid, maxCount := uint64(10), []byte("document-name"), 10
-	kv1 := NewKeyVersions(seqno, docid, int64(maxCount), 0)
-	kv2 := NewKeyVersions(seqno, docid, int64(maxCount), 0)
+	kv1 := NewKeyVersions(seqno, docid, maxCount, 0)
+	kv2 := NewKeyVersions(seqno, docid, maxCount, 0)
 	for i := 0; i < maxCount; i++ {
 		uuid := uint64(i * 10000)
 		kv1.AddUpsert(uuid, []byte("newkey"), []byte("oldkey"), nil)
@@ -53,7 +53,7 @@ func TestPayloadKeyVersions(t *testing.T) {
 		vbno, vbuuid := uint16(i), uint64(i*10)
 		vb := NewVbKeyVersions("default", vbno, vbuuid, 1000)
 		for j := 0; j < 10; j++ { // for 10 mutations
-			kv := NewKeyVersions(512 /*seqno*/, []byte("Bourne"), int64(nIndexes), 0)
+			kv := NewKeyVersions(512 /*seqno*/, []byte("Bourne"), nIndexes, 0)
 			kv.AddUpsert(uuids[0], keys[0], oldkeys[0], nil)
 			kv.AddUpsert(uuids[1], keys[1], oldkeys[1], nil)
 			kv.AddUpsert(uuids[2], keys[2], oldkeys[2], nil)
@@ -69,8 +69,8 @@ func TestPayloadKeyVersions(t *testing.T) {
 
 func BenchmarkKVEqual(b *testing.B) {
 	seqno, docid, maxCount := uint64(10), []byte("document-name"), 10
-	kv1 := NewKeyVersions(seqno, docid, int64(maxCount), 0)
-	kv2 := NewKeyVersions(seqno, docid, int64(maxCount), 0)
+	kv1 := NewKeyVersions(seqno, docid, maxCount, 0)
+	kv2 := NewKeyVersions(seqno, docid, maxCount, 0)
 	for i := 0; i < maxCount; i++ {
 		uuid := uint64(i * 10000)
 		kv1.AddUpsert(uuid, []byte("newkey"), []byte("oldkey"), nil)
